requirements.txt
GitPython==3.1.32
PyYAML==6.0.1

